+++ packages/cli/src/generators/backstage-integration/index.ts
@@ -4,7 +4,9 @@
 // https://opensource.org/licenses/MIT
 import {BaseGenerator} from '../../base-generator';
 import {BackstageIntegrationOptions} from '../../types';
-export default class BackstageIntegrationGenerator extends BaseGenerator<BackstageIntegrationOptions> {
+export default class BackstageIntegrationGenerator extends BaseGenerator<
+  BackstageIntegrationOptions
+> {
   cwd?: string;
   constructor(public args: string[], public opts: BackstageIntegrationOptions) {
     super(args, opts);