+++ packages/cli/src/generators/update/index.ts
@@ -16,7 +16,9 @@ const tempDeps = configJsonFile.config.templateDependencies;
 
 const packageJsonFile = 'package.json';
 
-export default class UpdateGenerator extends BaseUpdateGenerator<UpdateOptions> {
+export default class UpdateGenerator extends BaseUpdateGenerator<
+  UpdateOptions
+> {
   constructor(public args: string[], public opts: UpdateOptions) {
     super(args, opts);
   }