+++ services/authentication-service/src/modules/auth/google-login.controller.ts
@@ -74,8 +74,9 @@ export class GoogleLoginController {
   @get('/auth/google', {
     responses: {
       [STATUS_CODE.OK]: {
-        description:
-          'Google Token Response (Deprecated: Possible security issue if secret is passed via query params, please use the post endpoint)',
+        description: `Google Token Response,
+         (Deprecated: Possible security issue if secret is passed via query params, 
+          please use the post endpoint)`,
         content: {
           [CONTENT_TYPE.JSON]: {
             schema: {[X_TS_TYPE]: TokenResponse},
@@ -85,7 +86,9 @@ export class GoogleLoginController {
     },
   })
   /**
-   * @deprecated This method should not be used, possible security issue if secret is passed via query params, please use the post endpoint
+   * @deprecated
+   *  This method should not be used, possible security issue
+   *  if secret is passed via query params, please use the post endpoint
    */
   async loginViaGoogle(
     @param.query.string('client_id')