+++ services/authentication-service/src/modules/auth/index.ts
@@ -13,6 +13,7 @@ export * from './providers/resource-owner-verify.provider';
 export * from './providers/google-oauth2-verify.provider';
 export * from './providers/instagram-oauth2-verify.provider';
 export * from './providers/facebook-oauth-verify.provider';
+export * from './providers/cognito-oauth2-verify.provider';
 export * from './models/auth-user.model';
 export * from './providers/apple-oauth2-verify.provider';
 export * from './providers/otp-verify.provider';